requirements.txt
pandas
matplotlib
seaborn
scikit-learn
notebook
statsmodels
tensorflow
Django
djangorestframework
joblib